+++ changelog.md
@@ -266,6 +266,13 @@
 - Edit Admin Layout Navbar Components
 
 **v0.5.32**
-- February 25, 2025
+- February 26, 2025
 - Add Backend Controllers and Routes
 - Add Knowledgebase Feature: Completed Folder Features
+
+**v0.5.33**
+- February 27, 2025
+- Add Backend Controllers and Routes
+- Add Showdown JS
+- Add View Knowledgebase Page (Admin)
+- All Knowledgebase Feature for Admin: Completed